+++ src/main/kotlin/ru/yoomoney/gradle/plugins/moira/trigger/CollectMoiraTriggersTask.kt
@@ -1,14 +1,14 @@
 package ru.yoomoney.gradle.plugins.moira.trigger
 
 import org.gradle.api.DefaultTask
-import org.gradle.api.artifacts.Configuration
 import org.gradle.api.logging.Logging
+import org.gradle.api.plugins.JavaPluginExtension
 import org.gradle.api.tasks.Input
+import org.gradle.api.tasks.SourceSet
 import org.gradle.api.tasks.TaskAction
 import ru.yoomoney.gradle.plugins.moira.trigger.MoiraTriggerPlugin.Companion.TRIGGERS_FROM_ARTIFACT_DIR
 import ru.yoomoney.gradle.plugins.moira.trigger.collect.MoiraTriggersCollector
 import java.io.File
-
 import java.nio.file.Paths
 
 /**
@@ -27,23 +27,25 @@ open class CollectMoiraTriggersTask : DefaultTask() {
 
     @TaskAction
     fun collectMoiraTriggers() {
-        val artifactConfiguration = project.configurations.getByName("moiraFromArtifactCompile")
-        val dirConfiguration = project.configurations.getByName("moiraFromDirCompile")
+        val artifactSourceSet = project.extensions.getByType(JavaPluginExtension::class.java).sourceSets
+            .getByName("moiraFromArtifact")
+        val dirSourceSet = project.extensions.getByType(JavaPluginExtension::class.java).sourceSets
+            .getByName("moiraFromDir")
 
         val targetDir = Paths.get(project.projectDir.toString(), extension.dir).toFile()
-        targetDir.collectTriggers(dirConfiguration)
+        targetDir.collectTriggers(dirSourceSet)
 
         val targetDirArtifact = Paths.get(project.buildDir.toString(), TRIGGERS_FROM_ARTIFACT_DIR).toFile()
-        targetDirArtifact.collectTriggers(artifactConfiguration)
+        targetDirArtifact.collectTriggers(artifactSourceSet)
     }
 
-    private fun File.collectTriggers(configuration: Configuration) {
+    private fun File.collectTriggers(sourceSet: SourceSet) {
         if (!this.exists()) {
             log.lifecycle("Directory ${this.absolutePath} does not exist. Skip.")
             return
         }
 
-        val triggersCollector = MoiraTriggersCollector(configuration.files)
+        val triggersCollector = MoiraTriggersCollector(classPathFiles = sourceSet.compileClasspath.files)
         val collectedTriggers = triggersCollector.collect(this)
         collectedTriggers.forEach { (name, trigger) ->
             log.info("Collected trigger ({}): json={}", name, trigger.toJson().toString(4))
@@ -51,7 +53,6 @@ open class CollectMoiraTriggersTask : DefaultTask() {
     }
 
     companion object {
-
         private val log = Logging.getLogger(CollectMoiraTriggersTask::class.java)
     }
 }